Add plumbing for dynamic Clusters implementation.
This PR adds the plumbing required to pull cluster namespaces
dynamically. Actual code to perform data fetch and updates
will be added in follow on PRs. For the time being, if the
dynamic code path is invoked, an error is returned. Clients
must continue to specify namespace details in the coordinator
configuration.